requirements.txt
absl-py
astunparse
attrs
backcall
bleach
cachetools
certifi
chardet
click
colorama
cycler
decorator
defusedxml
entrypoints
Flask
Flask-Cors
gast
geojson
google-auth
google-auth-oauthlib
google-pasta

h5py
idna
importlib-metadata
ipykernel
ipython
ipython-genutils
ipywidgets
itsdangerous
jedi
Jinja2
joblib
jsonify
jsonschema
jupyter
jupyter-client
jupyter-console
jupyter-core
Keras-Preprocessing
kiwisolver
lxml
Markdown
MarkupSafe
matplotlib
mistune
nbconvert
nbformat
notebook
numpy
oauthlib
opt-einsum
packaging
pandas
pandas-datareader
pandocfilters
parso
pexpect
pickleshare
Pillow
prometheus-client
prompt-toolkit
protobuf
ptyprocess
pyasn1
pyasn1-modules
Pygments
pyparsing
pyrsistent
PySocks
python-dateutil
pytz

pyzmq
qtconsole
QtPy
requests
requests-oauthlib
rsa
scikit-learn
scipy
seaborn
Send2Trash
six
sklearn
tensorboard
tensorboard-plugin-wit
tensorflow
tensorflow-estimator
termcolor
terminado
testpath
threadpoolctl
tornado
traitlets
urllib3
wcwidth
webencodings
Werkzeug
widgetsnbextension
wincertstore
wrapt
zipp
